Modify the convolution time
My apology, In the previous program, the convolution time only consider the convolution for R, G, B layer. Now, I fix it including the send and receive the result from the slaves. So, we make a mistake in the last report. I try to upload the result on the GitHub.